feat: more

+++ frb_codegen/src/library/codegen/generator/wire/dart/spec_generator/misc/ty/dart_fn.rs
@@ -60,10 +60,10 @@ impl<'a> WireDartGeneratorMiscTrait for DartFnWireDartGenerator<'a> {
                 final serializer = SseSerializer(generalizedFrbRustBinding);
                 assert((rawOutput != null) ^ (rawError != null));
                 if (rawOutput != null) {{
-                    serializer.buffer.putInt32({action_normal});
+                    serializer.buffer.putUint8({action_normal});
                     sse_encode_{output_normal_safe_ident}(rawOutput.value, serializer);
                 }} else {{
-                    serializer.buffer.putInt32({action_error});
+                    serializer.buffer.putUint8({action_error});
                     sse_encode_{output_error_safe_ident}(rawError.value, serializer);
                 }}
                 final output = serializer.intoRaw();
